+++ lib/Calibre/CustomColumnTypeComment.php
@@ -15,16 +15,6 @@ class CustomColumnTypeComment extends CustomColumnType
         parent::__construct($pcustomId, self::CUSTOM_TYPE_COMMENT, $database);
     }
 
-    /**
-     * Get the name of the sqlite table for this column
-     *
-     * @return string
-     */
-    private function getTableName()
-    {
-        return "custom_column_{$this->customId}";
-    }
-
     public function getQuery($id)
     {
         global $config;
@@ -54,9 +44,14 @@ class CustomColumnTypeComment extends CustomColumnType
         return "<div>" . $value . "</div>"; // no htmlspecialchars, this is already HTML
     }
 
-    protected function getAllCustomValuesFromDatabase()
+    protected function getAllCustomValuesFromDatabase($n = -1)
     {
         return null;
+    }
+
+    public function getDistinctValueCount()
+    {
+        return 0;
     }
 
     public function getCustomByBook($book)