+++ lib/Calibre/CustomColumnTypeEnumeration.php
@@ -16,16 +16,6 @@ class CustomColumnTypeEnumeration extends CustomColumnType
     }
 
     /**
-     * Get the name of the sqlite table for this column
-     *
-     * @return string
-     */
-    private function getTableName()
-    {
-        return "custom_column_{$this->customId}";
-    }
-
-    /**
      * Get the name of the linking sqlite table for this column
      * (or NULL if there is no linktable)
      *
@@ -67,7 +57,8 @@ class CustomColumnTypeEnumeration extends CustomColumnType
 
     public function getCustom($id)
     {
-        $result = $this->getDb($this->databaseId)->prepare(str_format("SELECT id, value AS name FROM {0} WHERE id = ?", $this->getTableName()));
+        $query = str_format("SELECT id, value AS name FROM {0} WHERE id = ?", $this->getTableName());
+        $result = $this->getDb($this->databaseId)->prepare($query);
         $result->execute([$id]);
         if ($post = $result->fetchObject()) {
             return new CustomColumn($id, $post->name, $this);
@@ -75,12 +66,12 @@ class CustomColumnTypeEnumeration extends CustomColumnType
         return new CustomColumn(null, localize("customcolumn.enum.unknown"), $this);
     }
 
-    protected function getAllCustomValuesFromDatabase()
+    protected function getAllCustomValuesFromDatabase($n = -1)
     {
         $queryFormat = "SELECT {0}.id AS id, {0}.value AS name, count(*) AS count FROM {0}, {1} WHERE {0}.id = {1}.{2} GROUP BY {0}.id, {0}.value ORDER BY {0}.value";
         $query = str_format($queryFormat, $this->getTableName(), $this->getTableLinkName(), $this->getTableLinkColumn());
 
-        $result = $this->getDb($this->databaseId)->query($query);
+        $result = $this->getPaginatedResult($query, [], $n);
         $entryArray = [];
         while ($post = $result->fetchObject()) {
             $customcolumn = new CustomColumn($post->id, $post->name, $this);