+++ lib/Pages/PageCustomDetail.php
@@ -10,6 +10,9 @@ namespace SebLucas\Cops\Pages;
 
 use SebLucas\Cops\Calibre\BookList;
 use SebLucas\Cops\Calibre\CustomColumn;
+use SebLucas\Cops\Calibre\CustomColumnType;
+use SebLucas\Cops\Calibre\CustomColumnTypeDate;
+use SebLucas\Cops\Calibre\CustomColumnTypeInteger;
 
 class PageCustomDetail extends Page
 {
@@ -17,16 +20,41 @@ class PageCustomDetail extends Page
     {
         $customId = $this->request->get("custom", null);
         $custom = CustomColumn::createCustom($customId, $this->idGet, $this->getDatabaseId());
-        $this->getCustomEntries($custom);
         $this->idPage = $custom->getEntryId();
         $this->title = $custom->getTitle();
+        $this->getCustomEntries($custom->customColumnType);
         $this->parentTitle = $custom->customColumnType->getTitle();
         $this->parentUri = $custom->customColumnType->getUriAllCustoms();
     }
 
-    public function getCustomEntries($custom)
+    /**
+     * Summary of getCustomEntries
+     * @param CustomColumnType $columnType
+     * @return void
+     */
+    public function getCustomEntries($columnType)
     {
         $booklist = new BookList($this->request);
-        [$this->entryArray, $this->totalNumber] = $booklist->getBooksByCustom($custom, $this->idGet, $this->n);
+        if (empty($this->idGet)) {
+            if ($columnType instanceof CustomColumnTypeDate) {
+                // if we use $columnType::PAGE_DETAIL in PageAllCustoms, otherwise see PageAllCustoms
+                $year = $this->request->get("year", null, $columnType::GET_PATTERN);
+                if (!empty($year)) {
+                    [$this->entryArray, $this->totalNumber] = $booklist->getBooksByCustomYear($columnType, $year, $this->n);
+                    $this->title = $year;
+                    return;
+                }
+            }
+            if ($columnType instanceof CustomColumnTypeInteger) {
+                // if we use $columnType::PAGE_DETAIL in PageAllCustoms, otherwise see PageAllCustoms
+                $range = $this->request->get("range", null, $columnType::GET_PATTERN);
+                if (!empty($range)) {
+                    [$this->entryArray, $this->totalNumber] = $booklist->getBooksByCustomRange($columnType, $range, $this->n);
+                    $this->title = $range;
+                    return;
+                }
+            }
+        }
+        [$this->entryArray, $this->totalNumber] = $booklist->getBooksByCustom($columnType, $this->idGet, $this->n);
     }
 }