+++ test/customColumnsTest.php
@@ -731,7 +731,8 @@ class CustomColumnTest extends TestCase
         $this->assertEquals("cops:custom:12:year:2000", $currentPage->entryArray[0]->id);
         $this->assertEquals("2000", $currentPage->entryArray[0]->title);
         $this->assertEquals("4 books", $currentPage->entryArray[0]->content);
-        $this->assertEquals("phpunit?page=14&custom=12&year=2000", $currentPage->entryArray[0]->getNavLink());
+        // switched to using PAGE_DETAIL instead of PAGE_ALL
+        $this->assertEquals("phpunit?page=15&custom=12&year=2000", $currentPage->entryArray[0]->getNavLink());
         $this->assertEquals("cops:custom:12:year:2016", $currentPage->entryArray[1]->id);
 
         $config['cops_custom_date_split_year'] = '0';
@@ -771,6 +772,7 @@ class CustomColumnTest extends TestCase
         $request = new Request();
         $request->set('custom', 10);
 
+        $config['cops_custom_integer_split_range'] = '0';
         $currentPage = Page::getPage(Page::ALL_CUSTOMS, null, null, "1", $request);
         $currentPage->InitializeContent();
 
@@ -783,6 +785,20 @@ class CustomColumnTest extends TestCase
         $this->assertEquals("cops:custom:10:-1", $currentPage->entryArray[1]->id);
         $this->assertEquals("cops:custom:10:1", $currentPage->entryArray[2]->id);
         $this->assertEquals("cops:custom:10:2", $currentPage->entryArray[3]->id);
+
+        $config['cops_custom_integer_split_range'] = '4';
+        $currentPage = Page::getPage(Page::ALL_CUSTOMS, null, null, "1", $request);
+        $currentPage->InitializeContent();
+
+        $this->assertEquals("custom_08", $currentPage->title);
+        $this->assertCount(2, $currentPage->entryArray);
+        $this->assertEquals("cops:custom:10:range:-2--1", $currentPage->entryArray[0]->id);
+        $this->assertEquals("-2--1", $currentPage->entryArray[0]->title);
+        $this->assertEquals("4 books", $currentPage->entryArray[0]->content);
+        $this->assertEquals("phpunit?page=15&custom=10&range=-2--1", $currentPage->entryArray[0]->getNavLink());
+        $this->assertEquals("cops:custom:10:range:1-2", $currentPage->entryArray[1]->id);
+
+        $config['cops_custom_integer_split_range'] = '0';
     }
 
     public function testAllCustomsType09()